+++ alembic/versions/2c6458e01a73_first.py
@@ -1,0 +1,51 @@
+"""first
+
+Revision ID: 2c6458e01a73
+Revises: 
+Create Date: 2021-06-29 20:29:59.704520
+
+"""
+from alembic import op
+import sqlalchemy as sa
+import backend.database.models.types
+
+
+
+# revision identifiers, used by Alembic.
+revision = '2c6458e01a73'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('incidents',
+    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column('time_of_incident', sa.DateTime(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('officers',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('incident_id', sa.Integer(), nullable=False),
+    sa.Column('first_name', sa.Text(), nullable=True),
+    sa.Column('last_name', sa.Text(), nullable=True),
+    sa.Column('race', sa.Enum('UNKNOWN', 'WHITE', 'BLACK_AFRICAN_AMERICAN', 'AMERICAN_INDIAN_ALASKA_NATIVE', 'ASIAN', 'NATIVE_HAWAIIAN_PACIFIC_ISLANDER', name='race'), nullable=True),
+    sa.Column('gender', sa.Enum('UNKNOWN', 'MALE', 'FEMALE', 'TRANSGENDER', name='gender'), nullable=True),
+    sa.Column('appointed_date', sa.DateTime(), nullable=True),
+    sa.Column('badge', sa.Text(), nullable=True),
+    sa.Column('unit', sa.Text(), nullable=True),
+    sa.Column('rank', sa.Text(), nullable=True),
+    sa.Column('star', sa.Text(), nullable=True),
+    sa.Column('date_of_birth', sa.Date(), nullable=True),
+    sa.ForeignKeyConstraint(['incident_id'], ['incidents.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('officers')
+    op.drop_table('incidents')
+    # ### end Alembic commands ###
