+++ backend/routes/incidents.py
@@ -1,11 +1,9 @@
 from flask import Blueprint
 
 
-incident_routes = Blueprint(
-    "incident_routes", __name__, url_prefix="/incidents"
-)
+bp = Blueprint("incident_routes", __name__, url_prefix="/incidents")
 
 
-@incident_routes.route("", methods=["GET"])
+@bp.route("", methods=["GET"])
 def get_incidents():
     return ""
